Export BarLabel for users
