+++ packages/x-charts/src/BarChart/BarLabel/index.ts
@@ -1,5 +1,6 @@
-export { BarLabelComponent } from './BarLabel';
+export { BarLabel } from './BarLabel';
+export type { BarLabelProps } from './BarLabel';
 export { barLabelClasses, getBarLabelUtilityClass } from './barLabelClasses';
-export type { BarLabelSlotProps, BarLabelSlots, BarLabelProps } from './BarLabel';
-export type { BarLabelComponentProps, BarLabelOwnerState, BarItem, BarLabelContext } from './types';
+export type { BarLabelSlotProps, BarLabelSlots } from './BarLabelItem';
+export type { BarLabelOwnerState, BarItem, BarLabelContext } from './types';
 export type { BarLabelClasses, BarLabelClassKey } from './barLabelClasses';
