+++ packages/x-charts/src/BarChart/BarLabel/types.ts
@@ -10,10 +10,6 @@ export interface BarLabelOwnerState {
   classes?: Partial<BarLabelClasses>;
 }
 
-export type BarLabelComponentProps = Omit<React.SVGProps<SVGTextElement>, 'ref' | 'id'> & {
-  ownerState: BarLabelOwnerState;
-};
-
 export type BarItem = {
   /**
    * The series id of the bar.
